requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13

# Vector Stores
faiss-cpu==1.7.4
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# Evaluation Metrics
rouge-score==0.1.2
nltk==3.8.1

# LLM APIs
openai==1.10.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.7.2
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
